+++ dotnetcore/VocabularyAnalyser/Model/UserVocabulary.cs
@@ -7,19 +7,19 @@ namespace VocabularyAnalyser.Model
 {
     public class Language
     {
-        public int LanguageId { get; set; }
+        public Int64 LanguageId { get; set; }
         public string LanguageCode { get; set; }
         public string LanguageName { get; set; }
-        public int LanguageType { get; set; }
+        public Int64 LanguageType { get; set; }
     }
     public class MyVocabularyStatistics
     {
-        public int MyVocabularyStatisticsId { get; set; }
+        public Int64 MyVocabularyStatisticsId { get; set; }
         public string UserName { get; set; }
-        public int LanguageId { get; set; }
+        public Int64 LanguageId { get; set; }
         public virtual Language Language { get; set; }
         public double CurrentScore { get; set; }
-        public int TotalTestCount { get; set; }
+        public Int64 TotalTestCount { get; set; }
     }
     /// <brief>
     /// zh_CN,..., 不直接与WordAnalyser.Model.Language相关。通过独立工具为每一个用户
@@ -27,41 +27,41 @@ namespace VocabularyAnalyser.Model
     /// </brief>
     public class UserVocabulary
     {
-        public int UserVocabularyId { get; set; }
+        public Int64 UserVocabularyId { get; set; }
         public string UserName { get; set; }
         // index,
         public string WordUnicode { get; set; }
         // zh_CN,..., 不直接与WordAnalyser.Model.Language相关。通过独立工具为每一个用户
         // 创建或持续更新UserVocabulary。
-        public int LanguageId { get; set; }
+        public Int64 LanguageId { get; set; }
         public virtual Language Language { get; set; }
-        public int TestTimes { get; set; }
-        public int WrongTimes { get; set; }
+        public Int64 TestTimes { get; set; }
+        public Int64 WrongTimes { get; set; }
         /// <summary>
         /// 用于快速判断是否需要随机选择进行考核,一般小于等于0就是需要考核的。
         /// 测试一次成功+1，失败-1。
         /// </summary>
-        public int CriticalTimes { get; set; }
+        public Int64 CriticalTimes { get; set; }
     }
 
     public class MyVocabularyTest
     {
-        public int MyVocabularyTestId { get; set; }
+        public Int64 MyVocabularyTestId { get; set; }
         public string UserName { get; set; }
         [Display(Name = "考试题目数量:")]
-        public int TestWordCount { get; set; }
+        public Int64 TestWordCount { get; set; }
         public virtual List<VocabularyTestDetail> VocabularyTestDetails { get; set; }
         public DateTime TestTime { get; set; }
         public double Score { get; set; }
-        public int CorrectWordCount { get; set; }
-        public int LanguageId { get; set; }
+        public Int64 CorrectWordCount { get; set; }
+        public Int64 LanguageId { get; set; }
         public virtual Language Language { get; set; }
     }
 
     public class VocabularyTestDetail
     {
-        public int VocabularyTestDetailId { get; set; }
-        public int VocabularyTestId { get; set; }
+        public Int64 VocabularyTestDetailId { get; set; }
+        public Int64 VocabularyTestId { get; set; }
         public virtual MyVocabularyTest VocabularyTest { get; set; }
         // index,
         public string WordUnicode { get; set; }
