+++ dotnetcore/WordAnalyser/Model/WordResult.cs
@@ -11,8 +11,8 @@ namespace WordAnalyser.Model
     {
         public Int64 WordResultId { get; set; }
         public string WordLetter { get; set; }
-        public int BookId { get; set; }
+        public Int64 BookId { get; set; }
         public virtual Book Book { get; set; }
-        public int WordCount { get; set; }
+        public Int64 WordCount { get; set; }
     }
 }
